+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Showing an error when setting environment variables.
 - Indexes of the found items in the line.
 - Statistics output in the search engine for git projects.
+- Unpacking a git project.
 
 ### Fixed: